Add Updating of issues.
Adds in the ability for the gem to update a hit counter and add additional issues as notes  so that while it wont create a new update it will continue to provide error logs. So a better picture can be ascertained.

Also adds in the ability to reopen closed issues that have the same error.

Bumps version number

Adds a extra format field to the config. Old configs should still continue to work however this does add extra code to check if old or new config so this support will likeley be depreciated at some point to remove this code requirement.